requirements.txt
python-telegram-bot==13.15
pandas==1.3.5
requests==2.31.0
